Move BoardInitialCastle to board initial module
